Enforce input of type date, enhancing UX for browser compatible.
Note: the submitted date format is iso8601 YYYY-MM-DD (as we expect), 
even if the displayed date by the browser depends of the user's browser 
locale. Theses browsers then safely ignore input placeholder.

This is why we can't display the expected date format elsewhere in the 
interface: these users will see another format as the one we expect, and 
we don't want to mislead them.
The date format is only displayed as a placeholder, for browser 
non-compatible with input of type date.